[REFACTOR] Context provider를 App으로 이동
전체 label, issue, milestone fetch 결과는 다른 component에서도 사용합니다.
또한 이들끼리 깊은 관계를 이루기도 합니다.
그래서 Context의 범위를 App 전체로 올려놨습니다.
이제 App 아래에 있는 애들은 기본적으로 위 3개를 어디서든 접근 가능합니다.
예를 들어 이슈는
const {issues, setIssues} = useContext(IssueContext)
요런 식으로 가져올 수 있습니다.
