+++ client/src/App.js
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useState } from "react";
 import { Route, Switch } from "react-router-dom";
 import { createGlobalStyle } from 'styled-components';
 import Header from './components/header/header.js';
@@ -7,6 +7,7 @@ import Issue from "./components/issuelist/issuelist.js";
 import Label from './components/labellist/labellist.js';
 import NewIssue from "./components/newIssue/newIssue.js";
 import IssueDetail from "./components/issueDetail/issueDetail.js";
+import { IssueContext, LabelContext, MilestoneContext } from "./components/common/context.js";
 
 const ResetStyle = createGlobalStyle`
   body {
@@ -16,21 +17,29 @@ const ResetStyle = createGlobalStyle`
 `;
 
 const App = () => {
-    return (
-        <div>
-            <>
-                <ResetStyle />
-                <Header />
+  const [issues, setIssues] = useState([]);
+  const [labels, setLabels] = useState([]);
+  const [milestones, setMilestones] = useState([]);
+  
+  return (
+    <div>
+        <ResetStyle />
+        <Header />
+          <IssueContext.Provider value={{issues, setIssues}}>
+            <LabelContext.Provider value={{labels, setLabels}}>
+              <MilestoneContext.Provider value={{milestones, setMilestones}}>
                 <Switch>
                   <Route exact path="/issue/create" component={NewIssue}/>
                   <Route exact path="/issue/:id" component={IssueDetail} />
                   <Route exact path="/issue" component={Issue}/>
-                  <Route exact path='/label' component={Label}/>
                 </Switch>
-                <Route exact path="/" component={Login}/>
-            </>
-        </div>
-    )
+              </MilestoneContext.Provider>
+            </LabelContext.Provider>
+          </IssueContext.Provider>
+        <Route exact path='/label' component={Label}/>
+        <Route exact path="/" component={Login}/>
+    </div>
+  );
 }
 
 export default App;
