+++ client/src/components/common/context.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const IssueContext = React.createContext();
+const LabelContext = React.createContext();
+const MilestoneContext = React.createContext();
+
+export { IssueContext, LabelContext, MilestoneContext };
