--- client/src/components/issuelist/context.js
@@ -1,6 +1,0 @@
-import React from 'react';
-
-const FetchedDataContext = React.createContext();
-const IssueContext = React.createContext();
-
-export { FetchedDataContext, IssueContext };
