+++ client/src/components/issuelist/issuelist.js
@@ -1,11 +1,10 @@
-import React from "react";
+import React, { useContext } from "react";
 
 import { useIssues, useIssueLabels, useLabels, useMilestones } from './issueHook.js';
 
 import TopBar from '../topbar/topbar.js';
 import TabList from './tabList.js';
 import IssueItem from './issueItem.js';
-import { FetchedDataContext, IssueContext } from './context.js';
 
 const Issue = (props) => {
   const [issues, setIssues] = useIssues();
@@ -39,8 +38,7 @@ const Issue = (props) => {
     key={item.id} article={item} labels={labelMap[item.id]} onClickCheckbox={() => toggleIssueSelect(idx)}/>);
 
   return (
-    <IssueContext.Provider value={{issues, setIssues}}>
-    <FetchedDataContext.Provider value={{labels, milestones}}>
+    <div>
       <TopBar search={props.location.search} />
       <TabList
         labels={labels}
@@ -49,8 +47,7 @@ const Issue = (props) => {
       <div>
         {issueComponent}
       </div>
-    </FetchedDataContext.Provider>
-    </IssueContext.Provider>
+    </div>
   )
 }
